Add more options to the `unshard_checkpoint` function to help scale (#145)
I was getting worried about unsharding really big checkpoints like for
the 32B, which we'll need to do soon. The main issue at the moment is
that in order to unshard we need to load the entire model (or optimizer
state) in memory, which clearly isn't scalable. So I've added an option
to unshard the checkpoint into chunks of a given size which helps scale
because only a single chunk (which could be as small as a single tensor)
needs to load into memory at a time. Each chunk gets written to a unique
file. I think HuggingFace does something similar.

Note: this is not supported for optimizer state yet. But, speaking of
optimizer state, this PR also adds a function called `load_keys()` for
loading (and unsharding) specific keys from a checkpoint. So if you want
to inspect part of the optimizer state, you could use that function
without having to unshard the whole optimizer state.