Improve TP/CP APIs, add a fused linear loss (#185)
## Summary

- Fixes an error with TP.
- Moves all context parallel input data sharding logic to the
transformer model class so it's hidden from the user.
- Merges the loss function into the transformer LM head to make the TP
API cleaner and also to support a fused linear loss.
- Adds a low-memory fused linear loss option from Liger-Kernel.

## Breaking changes

- Removed the options `fused_loss` and `compile_loss` from
`TransformerTrainModule`. Since the loss function is now part of the
model, it will be compiled when `compile_model` is `True`. You can
change the loss implementation via
`--model.lm_head.loss_implementation`.
- The `TransformerPipelineTrainModule` is just broken, so I've removed
it from the public API for now.