Bump ai2-olmo-eval==0.7.0 (in-loop Minerva, GSM, HumanEval, MBPP) (#204)
Two changes:

1. Increment `ai2-olmo-eval==0.7.0`:
https://github.com/allenai/OLMo-in-loop-evals/releases/tag/v0.7.0

This allows us to use the following new in-loop eval config options:

```python
DownstreamEvaluatorCallbackConfig(
  tasks=[
    "gsm8k_gold_bpb_5shot",
    "minerva_math_algebra_gold_bpb_0shot",
    "minerva_math_counting_and_probability_gold_bpb_0shot",
    "minerva_math_geometry_gold_bpb_0shot",
    "minerva_math_intermediate_algebra_gold_bpb_0shot",
    "minerva_math_number_theory_gold_bpb_0shot",
    "minerva_math_prealgebra_gold_bpb_0shot",
    "minerva_math_precalculus_gold_bpb_0shot",
    "codex_humaneval_gold_bpb_0shot",
    "codex_mbpp_gold_bpb_0shot",
    ...
```

2. Fix in-loop eval bug in V2.

I believe in V2 the dataloader was changed for evals. The new dataloader
cut-off the last example used as input, so the resulting logprobs look
like this (e.g., `csqa_val_rc_5shot`):

```python
{
    0: {0: -2.6875, 2: -1.390625, 4: -2.921875, 1: -1.4921875, 3: -2.96875},
    1: {1: -2.1875, 3: -2.5, 0: -1.9375, 2: -1.9765625, 4: -2.5625},
    2: {0: -1.9921875, 2: -1.8671875, 4: -3.0625, 1: -1.390625, 3: -1.75},
    ...
    1176: {0: -1.578125, 2: -2.3125, 4: -2.34375, 1: -2.046875, 3: -1.28125},
    1177: {1: -2.109375, 0: -1.6875, 2: -1.4921875}, # <- Should have 5 entries, because each question has 5 responses!
}
```

This specifically breaks eval suites where the label of the last
question does not fit in the batch. So, I changed the eval dataloader to
use `drop_last=False`, which matches the `DistributedSampler` in the
same file.