fix in-loop normalization with v2 (#243)
Merge and update in-loop evals to match
https://github.com/allenai/OLMo-in-loop-evals/pull/6

Description from the `ai2-olmo-eval==0.7.1` PR:
> our current in-loop does not match the OLMES standard (in our
downstream evals), [the downstream normalization is
here](https://github.com/allenai/oe-eval-internal/blob/main/oe_eval/metrics/metric.py#L229-L231).
It appears the -1 term is an artifact from when we originally
implemented in-loop evals.

This will keep the original version of the metrics and add a
version with the fixed normalization (`_v2`)

---------

Co-authored-by: Dirk Groeneveld <dirkg@allenai.org>