Update images to stable 2.7.0, use CUDA 12.8 by default (#280)
Also makes a change so we always add the build date to the image tag, so
we never override existing images.

I tested the new images I just build on Jupiter and Augusta with a 7B.
They seemed to work fine, other than some new version of a dependency
messing with the logging configuration. I never tracked down which
dependency was doing that, but I found out we can enforce our own
configuration by adding `force=True` to our call to
`logging.basicConfig(...)`.