Make async bookkeeping more robust (#305)
Adds a queue for tracking bookkeeping operations and the ability to
cancel queued operations of the same type when submitting a new one:
`trainer.run_bookkeeping_op(...., cancel_in_progress=True)`. The trainer
will throw a warning when attempting to cancel a queued op that's
currently running, which can be a good indication that the op is causing
a backlog in the queue. The trainer will also throw a warning if any
bookkeeping op takes longer than 30 seconds by default.