Memory budget strategy for activation checkpointing (#297)
See https://pytorch.org/blog/activation-checkpointing-techniques/ for
more details, but essentially this is an easy way to try to enable
selective activation checkpointing without fiddling with a bunch of
different options to try to make it fast but stay within your GPU memory
allowance.


![image](https://github.com/user-attachments/assets/5e17af03-aa43-489e-b30e-471ee3025c7e)

> We observe a 50% memory reduction by recomputing only pointwise ops,
with a steady drop-off as you recompute more and more of your matmuls.
Attention is the most expensive, so you tend to want to recompute those
last.