fix initialization (again) (#319)
While fixing one thing, #307 introduced a different issue with HSDP.
Unlike DDP, PyTorch's `fully_shard()` (FSDP) wrapper doesn't synchronize
to make sure parameters are consistent across the process group. So when
you init your rank-local tensors with different seeds, you end up with
inconsistent parameters across replica groups. That means replica 1 has
completely different parameters than replica 2, and so on. When we
restarted the bad run, this actually helped it out because now the
replicas all ended up with the same parameters (the params from the
first replica, because that's what was saved to the checkpoint).

This PR reverts #307 more or less, and introduces a different way of
initializing DTensors in general such that:
1. We don't run into this bug with HSDP.
2. We get the same full tensors after initialization regardless of world
size.

I tested this end-to-end with the exact same OLMo3-1B settings where we
observed the issue in the first place. You can find the runs here named
`OLMo3-1B-swafix-initfix`:
https://wandb.ai/ai2-llm/olmo3/workspace?nw=2kjfyryx4wm

![image](https://github.com/user-attachments/assets/7beff958-2fd8-4e34-be08-38fb542eea45)

Interestingly we also see lower peak GPU memory usage with this method.

![image](https://github.com/user-attachments/assets/5ef5e7c3-5e4a-45cb-b56e-cca019bd93d8)