Add `TrainModule.pre_train()`, fix pain point with BZ scheduler (#332)
We have a pain point with the `BatchSizeScheduler` where a batch size
check will fail (raising a `RuntimeError`) when restarting a run after a
batch size increase with a bigger world size than would have been
allowed with the previous batch size. As a result we've had to change
parts of training script before relaunching, which is annoying, easy to
forget, and error-prone.

The issue happens because the batch size check occurs before the
`BatchSizeScheduler` updates the new batch size on
`.post_checkpoint_loaded()`. So to fix this, this PR postpones the batch
size check until right before training starts by moving it to a new
`TrainModule` method `.pre_train()`, which runs right after
`Callback.pre_train()`.