Improve distributed error handling (#380)
I've seen a few scenarios recently where an error occurs before training
begins, either during loading a checkpoint or in the `pre_train()`
method of a callback, and instead of failing immediately the process
will hang for the duration of the distributed timeout. It turns out
that's because we were setting up our training scripts like this:

```python
try:
    train()
finally:
    teardown_distributed_environment()
```

So we'd end up trying to teardown the distributed process group from one
rank (the first rank to raise an exception) while the others are doing
something else, leading to the hang. A fix for this is to only call the
teardown function when the training loop completes successfully:

```python
train()
teardown_distributed_environment()
```

lI tested this out with a callback that raises an exception on
`pre_train()` and this seems to work well. The job ended within a few
seconds after the error as opposed to 30 mins (the NCCL timeout).