Add attention backend abstraction and integrate transformer engine's attention (#384)
# Overview

This PR adds a new abstraction called `AttentionBackend` that
encapsulates an SDPA kernel, making it easier to add new attention
kernels. It also adds a TransformerEngine attention backend, which wraps
TE's
[DotProductAttention](https://docs.nvidia.com/deeplearning/transformer-engine/user-guide/api/pytorch.html#transformer_engine.pytorch.DotProductAttention).
This gives us a nice speed-up even over flash-attn, both with and
without context parallelism.

The attention backed is now configured with the `backend` argument to
the `Attention` and `AttentionConfig` classes. Valid values are
currently "torch", "flash", and "te". The `use_flash` argument is
deprecated in favor of `backend`. And if neither is specified, we'll
choose a compatible backend based on other configuration options.

So to use the TransformerEngine attention backend, just add the override
`--model.block.attention.backend=te`.

# For reviewers

So far the commit history is clean and can be reviewed one-by-one, if
that makes it easier.
- 4e622610fa38515ca3588c8c832b65eabf4f94e0 refactors the attention
module to use the new `AttentionBackend` abstraction.
- 25256f488c34f523476b42dd262572bee9cfa279 adds TE to our Docker builds.
- 73195807511d43cb04c9abab0b942552d3394dfa adds a TE attention backend.