clean up checkpointing code to minimize distributed communication (#428)
We're consistently running into failures while checkpointing on Augusta
(this is synchronous checkpointing since GLOO has been an issue), and
the point of failure is often a `barrier` right before we actually write
the checkpoint. We're still investigating _why_ we're failing at a
barrier, but we hope that reducing the number of barriers (and
distributed communication in general) will reduce the probability of
failure.

So this PR:
- Reduces the number of `barrier` calls prior to writing the checkpoint.
- Reduces the amount of communication required for subsequent trainer
checkpoints by caching the global save plan.